% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_elements.R
\name{boostrap_table}
\alias{boostrap_table}
\title{convert a data.frame to a boostrap table}
\usage{
boostrap_table(x, class = "responsive")
}
\arguments{
\item{x}{a data.table}

\item{class}{`div` class}
}
\description{
convert a data.frame to a boostrap table
}
\note{
a wrapper around knitr::kable
}
\examples{
boostrap_table( data.table::data.table(x = 1, y = 'a') )
}
