% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleaner}
\alias{cleaner}
\title{Cleans a data table
Removes NA rows, replaces 'NA' with NA}
\usage{
cleaner(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
Cleans a data table
Removes NA rows, replaces 'NA' with NA
}
