% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{DataEntry}
\alias{DataEntry}
\alias{.onLoad}
\alias{DataEntry-package}
\title{Data entry 'in-browser' interfaces}
\usage{
.onLoad(libname, pkgname)
}
\description{
Data-entry front-end; Mysql/MariaDB backend
}
