% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dataentry_deps}
\alias{dataentry_deps}
\title{DataEntry HTML dependencies}
\usage{
dataentry_deps()
}
\value{
HTML dependencies used by DataEntry Shiny apps.
}
\description{
DataEntry HTML dependencies
}
