% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-append_rows.R,
%   R/app-server-edit_rcode.R, R/app-server-edit_table.R, R/app-server.R
\name{server_append_rows}
\alias{server_append_rows}
\alias{server_edit_rcode}
\alias{server_edit_table}
\alias{dataentry_servers}
\title{Data entry server modules}
\usage{
server_append_rows(input, output, session)

server_edit_rcode(input, output, session)

server_edit_table(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
Called for side effects.

Called for side effects.
}
\description{
Server modules for appending rows, editing database tables, and editing
inspector definitions.
}
\note{
See \code{global.R} in \code{inst/UI/newData} for required variables to set.

See \code{global.R} in \code{inst/UI/editInspector} for required variables to set.

See \code{global.R} in \code{inst/UI/editData} for required variables to set.
}
