% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ui.R
\name{ui_append_rows}
\alias{ui_append_rows}
\alias{ui_edit_table}
\alias{ui_edit_inspectors}
\title{Data entry UI}
\usage{
ui_append_rows(table_name = "Table Name")

ui_edit_table(table_name = "Table Name")

ui_edit_inspectors(table_name = "inspectors")
}
\arguments{
\item{table_name}{Label shown in the data-entry menu.}
}
\value{
A Shiny UI object.
}
\description{
Data entry UI
}
