% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI-s.R
\name{DataEntryDashboard}
\alias{DataEntryDashboard}
\title{DataEntryMaterialFloating}
\usage{
DataEntryDashboard(tableName, skin = "yellow", height = "100\%",
  width = "100\%")
}
\arguments{
\item{tableName}{}

\item{skin}{default to 'yellow'}

\item{height}{default to 100}

\item{width}{default to 100}
}
\description{
DataEntryMaterialFloating
}
\examples{
\dontrun{
frm = function() { H<<- data.table(a = 1:10, b = rep(NA, 10)) }
shinyApp(
 onStart= frm,
 ui = DataEntryDashboard('table name'), 
 server = function(input, output,session) { 
 output$table  <- renderRHandsontable({
  rhandsontable(H)
 } ) } )
 }
}
