% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI-s.R
\name{DataEntryMaterialFloating}
\alias{DataEntryMaterialFloating}
\title{DataEntryMaterialFloating}
\usage{
DataEntryMaterialFloating(tableName, skin = "yellow", height = "100\%",
  width = "100\%")
}
\arguments{
\item{tableName}{}

\item{skin}{default to 'yellow'}

\item{height}{default to 100}

\item{width}{default to 100}
}
\description{
DataEntryMaterialFloating
}
