% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{db_get}
\alias{db_get}
\title{Run a MariaDB query and return a data.table}
\usage{
db_get(query, ...)
}
\arguments{
\item{query}{SQL query passed to \code{DBI::dbGetQuery()}.}

\item{...}{Additional arguments passed to \code{db_con()}.}
}
\value{
A data.table.
}
\description{
Run a MariaDB query and return a data.table
}
