% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.R
\name{dropDownNavPage}
\alias{dropDownNavPage}
\title{shiny ui based on shinyWidgets}
\usage{
dropDownNavPage(tableName = "Table Name")
}
\description{
shiny ui based on shinyWidgets
}
\examples{
if (interactive()) {
require(DataEntry)
shinyApp(
 ui = dropDownNavPage(), 
 server = function(input, output) { 
     output$table <- renderRHandsontable( 
         rhandsontable(matrix(as.integer(NA), 
nrow = 30, ncol = 20) |> data.table() ) )
 } )


}

}
