% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emptyFrame}
\alias{emptyFrame}
\title{emptyFrame
emptyFrame used by handsontable}
\usage{
emptyFrame(table, n = 10, excludeColumns = "pk", preFilled, colorder)
}
\arguments{
\item{table}{db table}

\item{n}{n empty lines}

\item{excludeColumns}{default 'pk'}

\item{preFilled}{a named list eg. list(datetime_ = as.character(Sys.Date()))}

\item{colorder}{column order. see data.table::setcolorder}
}
\description{
emptyFrame
emptyFrame used by handsontable
}
