% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_rowid}
\alias{ensure_rowid}
\title{Ensure row ids
Add a \code{rowid} column if missing.}
\usage{
ensure_rowid(x)
}
\arguments{
\item{x}{A \code{data.table}.}
}
\value{
\code{x}, invisibly.
}
\description{
Ensure row ids
Add a \code{rowid} column if missing.
}
