% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{grand_n}
\alias{grand_n}
\title{grand_n
N rows in a db table}
\usage{
grand_n(user, host, db, pwd, table)
}
\arguments{
\item{user}{db user}

\item{host}{db host}

\item{db}{db name}

\item{pwd}{pwd}

\item{table}{db table}
}
\description{
grand_n
N rows in a db table
}
