% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{hhmm_validator}
\alias{hhmm_validator}
\title{HH:MM validator}
\usage{
hhmm_validator(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
HH:MM validator
}
\examples{
x = data.table(v1 = c('02:04' , '16:56', '23:59'  ), v2 = c('24:04' , NA, '23:59'  ) )
hhmm_validator(x)
}

