% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{inspector}
\alias{inspector}
\title{data inspector}
\usage{
inspector(x)
}
\arguments{
\item{x}{a data.table with its (S3) class extended by the database table name (see server.R)}
}
\description{
Data police
}
\details{
Inspectors are usually a collection of validators
}
