% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-inspector.R
\name{inspector_loader}
\alias{inspector_loader}
\title{Inspector loader}
\usage{
inspector_loader(table_name, ...)
}
\arguments{
\item{table_name}{Database table name.}

\item{...}{Additional arguments passed to \code{\link[=db_get]{db_get()}}.}
}
\value{
A function with signature \verb{function(x, ...)}.
}
\description{
Inspector loader
}
