% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{interval_validator}
\alias{interval_validator}
\title{interval (lower, upper)}
\usage{
interval_validator(x, v)
}
\arguments{
\item{x}{a data.table}

\item{v}{a data.table with variable, lq, uq columns}
}
\description{
interval (lower, upper)
}
\examples{
x = data.table(v1 = runif(5)  , v2 = runif(5) )
v = data.table(variable = c('v1', 'v2'), lq = c(-1, 0.2), uq = c(.7, 0.5) )
interval_validator(x)
}

