% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is.na_validator}
\alias{is.na_validator}
\title{is.na_validator}
\usage{
is.na_validator(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
is.na_validator
}
\examples{
x = data.table(v1 = c(1,2, NA, NA), v2  = c(1,2, NA, NA) )
is.na_validator(x)
}

