% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-js.R
\name{jquery_change_by_id}
\alias{jquery_change_by_id}
\title{find and replace given \code{div} id.}
\usage{
jquery_change_by_id(divid, newtext)
}
\arguments{
\item{divid}{div id}

\item{newtext}{new text entry}
}
\description{
find and replace given \code{div} id.
}
\note{
used inside shiny apps
}
