% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js_before_unload}
\alias{js_before_unload}
\title{prevent page exit}
\usage{
js_before_unload(msg = "Are you sure you want to exit the page?")
}
\arguments{
\item{msg}{msg}
}
\description{
prevent page exit
}
\note{
used inside shiny apps
}
