% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js_insertMySQLTimeStamp}
\alias{js_insertMySQLTimeStamp}
\title{translate forward slash to a Mysql timestamp}
\usage{
js_insertMySQLTimeStamp()
}
\description{
translate forward slash to a Mysql timestamp
}
\note{
used inside shiny apps
}
