% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meltall}
\alias{meltall}
\title{melt all columns in a data.table}
\usage{
meltall(x, na.rm = TRUE)
}
\arguments{
\item{x}{a data.table}

\item{na.rm}{TRUE by default}
}
\description{
melt all columns in a data.table
}
