% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{nchar_validator}
\alias{nchar_validator}
\title{N char}
\usage{
nchar_validator(x, v)
}
\arguments{
\item{x}{a data.table}

\item{v}{a data.table with variable and N (n char)}
}
\description{
N char
}
\examples{
x = data.table(v1 = c('x', 'xy', 'x')  , v2 = c('xx', 'x', 'xxx')  )
v = data.table(variable = c('v1', 'v2'), n = c(1, 2) )
interval_validator(x)
}

