% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{POSIXct_validator}
\alias{POSIXct_validator}
\title{POSIXct_validator}
\usage{
POSIXct_validator(x)
}
\arguments{
\item{x}{a data.table}
}
\description{
POSIXct_validator
}
\examples{
x = data.table(v1 = c(NA, '2016-11-23 10:23:00', as.character(Sys.time() - 3600*24*10 )  ),
               v2 = c('2016-11-23 25:23', '2016-11-23 30:80', as.character(Sys.time() + 100000)  ) )
POSIXct_validator(x)
}

