% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_append_rows.R
\name{server_append_rows}
\alias{server_append_rows}
\title{server_append_rows}
\usage{
server_append_rows(input, output, session)
}
\description{
server_append_rows
}
\note{
uitable, comments are hardwired and should be defined in global.R
inspectors are loaded with \code{\link[=inspector_loader]{inspector_loader()}}
}
