% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-edit_inspectors.R
\name{server_edit_inspectors}
\alias{server_edit_inspectors}
\title{server_edit_inspectors}
\usage{
server_edit_inspectors(input, output, session)
}
\description{
server_edit_inspectors
}
\note{
see global.R in inst/UI/editInspector for required variables to set.
}
