% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_edit_table.R
\name{server_edit_table}
\alias{server_edit_table}
\title{server_edit_table}
\usage{
server_edit_table(input, output, session)
}
\description{
server_edit_table
}
\note{
uitable, comments, backupdir are hardwired and should be defined in global.R
}
