% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strp_date_or_time}
\alias{strp_date_or_time}
\title{strp datetime or time (Mysql compatible)}
\usage{
strp_date_or_time(x)
}
\arguments{
\item{x}{strip datetime or date}
}
\description{
strp datetime or time (Mysql compatible)
}
\examples{
x = c(Sys.Date() \%>\% as.character, Sys.time()\%>\% as.character )
strp_date_or_time(x)
}
