% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.R
\name{ui_vnavbar}
\alias{ui_vnavbar}
\title{shiny ui with a vertical navbar and a rHandsontable}
\usage{
ui_vnavbar()
}
\description{
shiny ui with a vertical navbar and a rHandsontable
}
\note{
see praise package
}
\examples{
if (interactive()) {
tableName = 'TABLE X'
shinyApp(
	ui = ui_vnavbar(), 
	server = function(input, output) { 
		output$table <- renderRHandsontable( 
			rhandsontable(iris)
			)
	} )


}

}
