% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_navbar.R
\name{useNavbar}
\alias{useNavbar}
\title{Initialize the bootstrap navbar}
\usage{
useNavbar()
}
\value{
The HTML tags to put into the \code{<head>} of the HTML file.
}
\description{
Call this function once from inside Shiny UI
}
