% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vnavbarPage}
\alias{vnavbarPage}
\title{shiny ui with a vertical navbar and a rHandsontable}
\usage{
vnavbarPage(tableName = "Table Name")
}
\description{
shiny ui with a vertical navbar and a rHandsontable
}
\examples{
if (interactive()) {
require(DataEntry)
shinyApp(
 ui = vnavbarPage(), 
 server = function(input, output) { 
     output$table <- renderRHandsontable( 
         rhandsontable(matrix(as.integer(NA), nrow = 30, ncol = 20) \%>\% data.table) )
 } )


}
}
