% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{boostrap_table}
\alias{boostrap_table}
\title{data.table to html (boostrap style)}
\usage{
boostrap_table(x, class = "responsive")
}
\arguments{
\item{x}{a data.table}
}
\description{
data.table to html (boostrap style)
}
\examples{
boostrap_table( data.table(x = 1, y = 'a') )
}
