% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVER_elements.R
\name{column_comment}
\alias{column_comment}
\title{column_comment 
db table comments}
\usage{
column_comment(user, host, db, pwd, table, excludeColumns = "pk")
}
\arguments{
\item{user}{db user}

\item{host}{db host}

\item{db}{db name}

\item{pwd}{pwd}

\item{table}{db table}

\item{excludeColumns}{default 'pk'}
}
\description{
column_comment 
db table comments
}
