% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{column_comment}
\alias{column_comment}
\title{column_comment
db table comments}
\usage{
column_comment(user, host, db, pwd, table, excludeColumns = "pk")
}
\arguments{
\item{user}{db user}

\item{host}{db host}

\item{db}{db name}

\item{pwd}{pwd}

\item{table}{db table}

\item{excludeColumns}{default 'pk'}
}
\value{
a data.frame with two fields: Column and description
}
\description{
column_comment
db table comments
}
\note{
this is just a convenience function around a "Select .. from information_schema" query.
}
