% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_queries.R
\name{column_comment}
\alias{column_comment}
\title{select mariadb column's comment}
\usage{
column_comment(table, db, user, host)
}
\arguments{
\item{table}{table name}

\item{user}{db user}

\item{host}{host}
}
\description{
select mariadb column's comment
}
