\name{DataEntry-package}
\alias{DataEntry-package}
\alias{DataEntry}
\docType{package}
\title{
\packageTitle{DataEntry}
}
\description{
\packageDescription{DataEntry}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DataEntry}
\packageIndices{DataEntry}
}
\author{
\packageAuthor{DataEntry}

Maintainer: \packageMaintainer{DataEntry}
}
