% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{DataEntry-package}
\alias{DataEntry}
\alias{DataEntry-package}
\title{Data entry 'in-browser' interfaces}
\description{
Data-entry front-end; MySQL/MariaDB backend. Flexible support for front-end data validation.
}
\keyword{internal}
