% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEntry-package.R
\docType{package}
\name{DataEntry-package}
\alias{DataEntry}
\alias{DataEntry-package}
\title{Data entry 'in-browser' interfaces}
\description{
Data-entry front-end with a MariaDB backend. Flexible support for front-end data validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ornitho-logics.github.io/DataEntry/}
  \item Report bugs at \url{https://github.com/ornitho-logics/DataEntry/issues}
}

}
\keyword{internal}
