% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{DataEntry}
\alias{DataEntry-package}
\alias{DataEntry}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{  Data entry 'in-browser' interfaces
}\if{html}{\out{</div>}}}
\description{
Data-entry front-end; Mysql/MariaDB backend. Flexible support for front-end data validation.
}
