% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{db_con}
\alias{db_con}
\title{Create a MariaDB connection}
\usage{
db_con(.cnf = get("cnf_path", envir = .GlobalEnv))
}
\arguments{
\item{host}{Database host.}

\item{user}{Database user.}

\item{pwd}{Database password.}

\item{db}{Database name.}
}
\value{
A MariaDB connection.
}
\description{
Uses credentials from the global environment by default.
}
