% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{db_con}
\alias{db_con}
\title{Create a MariaDB connection}
\usage{
db_con(.cnf = NULL, group = "DataEntry")
}
\arguments{
\item{.cnf}{Path to a MariaDB option file.}

\item{group}{Option group in \code{.cnf}.}
}
\value{
A MariaDB connection.
}
\description{
Creates a MariaDB connection from an option file.
}
\details{
If \code{.cnf} is not supplied, \code{db_con()} first looks for \code{cnf_path} in the global
environment, then for the \code{DATAENTRY_CNF} environment variable. The option
file is read with the \code{DataEntry} group by default.
}
