% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emptyFrame}
\alias{emptyFrame}
\title{emptyFrame
emptyFrame used by handsontable}
\usage{
emptyFrame(
  user,
  host,
  db,
  pwd,
  table,
  n = 10,
  excludeColumns = "pk",
  preFilled,
  colorder
)
}
\arguments{
\item{user}{db user}

\item{host}{db host}

\item{db}{db name}

\item{pwd}{pwd}

\item{table}{db table}

\item{n}{n empty lines}

\item{excludeColumns}{default 'pk'}

\item{preFilled}{a named list eg. list(datetime_ = as.character(Sys.Date()))}

\item{colorder}{column order. see data.table::setcolorder}
}
\description{
emptyFrame
emptyFrame used by handsontable
}
\examples{
\dontrun{
emptyFrame(user = "testuser", pwd = "testuser", host = "127.0.0.1", db = "tests", table = "data_entry")
emptyFrame(
  user = "testuser", pwd = "testuser", host = "127.0.0.1", db = "tests", table = "data_entry",
  preFilled = list(datetime_ = as.character(Sys.Date()))
)
emptyFrame(
  user = "testuser", pwd = "testuser", host = "127.0.0.1", db = "tests", table = "data_entry",
  colorder = c("ID", "sex", "nest")
)
}

}
