% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{emptyFrame}
\alias{emptyFrame}
\title{emptyFrame 
emptyFrame used by handsontable}
\usage{
emptyFrame(user, host, db, table, n = 10, excludeColumns = "pk", preFilled)
}
\arguments{
\item{user}{db user}

\item{host}{db host}

\item{db}{db name}

\item{table}{db table}

\item{n}{n empty lines}

\item{excludeColumns}{default 'pk'}

\item{preFilled}{a named list eg. list(datetime_ = as.character(Sys.Date()))}
}
\description{
emptyFrame 
emptyFrame used by handsontable
}
