% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset_queries.R
\name{grand_n}
\alias{grand_n}
\title{table size (nrow)}
\usage{
grand_n(table, db, user, host)
}
\arguments{
\item{table}{table name}

\item{user}{db user}

\item{host}{host}
}
\description{
table size (nrow)
}

