% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\name{inspector}
\alias{inspector}
\title{data inspector}
\usage{
inspector(sql, x, user, db, host)
}
\arguments{
\item{sql}{sql string returning the script stored in db}

\item{x}{data to be validated.}
}
\value{
a data.table
}
\description{
an inspector runs all the validators in the script.
}
\examples{
inspector('select script from validators where table_name = "test_tbl"', 
         x, 'mihai', 'test', '127.0.0.1')
}
