% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_elements.R, R/js.R
\name{jquery_change_by_id}
\alias{jquery_change_by_id}
\title{find and replace given `div` id.}
\usage{
jquery_change_by_id(divid, newtext)

jquery_change_by_id(divid, newtext)
}
\arguments{
\item{divid}{div id}

\item{newtext}{new text entry}

\item{divid}{div id}

\item{newtext}{new text entry}
}
\description{
find and replace given `div` id.

find and replace given `div` id.
}
\note{
used inside shiny apps

used inside shiny apps
}
