% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js_hot_tippy_header}
\alias{js_hot_tippy_header}
\title{add tooltips to a handsontable}
\usage{
js_hot_tippy_header(x, tippy_column)
}
\arguments{
\item{x}{A data frame with two columns, one colum containing the fields of the table
and one column containing the description of the fields.}

\item{tippy_column}{Character vector. the name of the column in x containing the description of the columns.}
}
\description{
add tooltips to a handsontable
}
\note{
used within rhandsontable: e.g. rhandsontable(afterGetColHeader = js_hot_tippy_header(comments, "description"))
}
\seealso{
\code{\link[=column_comment]{column_comment()}}
}
