% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-server-append_rows.R
\name{server_append_rows}
\alias{server_append_rows}
\title{server_append_rows}
\usage{
server_append_rows(input, output, session)
}
\description{
server_append_rows
}
\note{
see global.R in inst/UI/newData for required variables to set.
}
