% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVER.R
\name{server_newData}
\alias{server_newData}
\title{Title}
\usage{
server_newData(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Title
}
\note{
inspector, uitable, comments, describeTable, getDataSummary are defined in global.R
}
