% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVER.R
\name{server_newData}
\alias{server_newData}
\title{Shiny server function}
\usage{
server_newData(input, output, session)
}
\arguments{
\item{input}{Shiny server}

\item{output}{Shiny server}

\item{session}{Shiny server}
}
\description{
Shiny server function
}
\note{
package, uitable, comments, describeTable, getDataSummary are defined in global.R
      inspectors are loaded with  DataEntry.validation::inspector_loader()
}
