% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVER.R
\name{server_newData_dropDownNavPage}
\alias{server_newData_dropDownNavPage}
\title{Shiny server function}
\usage{
server_newData_dropDownNavPage(input, output, session)
}
\arguments{
\item{input}{Shiny server}

\item{output}{Shiny server}

\item{session}{Shiny server}
}
\description{
Shiny server function
}
\note{
package, uitable, comments, describeTable, getDataSummary are hardwired and should be defined in global.R
inspectors are loaded with  \code{\link[DataEntry.validation:inspector_loader]{DataEntry.validation::inspector_loader()}}
}
