% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_elements.R
\name{useNavbar}
\alias{useNavbar}
\title{Initialize the bootstrap navbar
Call this function once from inside Shiny UI}
\usage{
useNavbar()
}
\value{
The HTML tags to put into the \code{<head>} of the HTML file.
}
\description{
Initialize the bootstrap navbar
Call this function once from inside Shiny UI
}
